LSP/requirements.txt
absl-py==1.4.0
albumentations==0.5.2
appdirs==1.4.4
attrs==22.2.0
av==8.0.3
audioread==3.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
cachetools==5.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cycler==0.11.0
debugpy==1.6.5
decorator==4.4.2
docker-pycreds==0.4.0
dominate==2.7.0
entrypoints==0.4
flatbuffers==23.1.4
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
h5py==3.7.0
idna==3.4
imageio==2.24.0
imageio-ffmpeg==0.4.8
imgaug==0.4.0
importlib-metadata==6.0.0
ipykernel==6.16.2
ipython==7.34.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mediapipe==0.9.0.1
moviepy==1.0.3
nest-asyncio==1.5.6
networkx==2.6.3
numba==0.56.4
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.5.5.64
opencv-python-headless==4.6.0.66
packaging==23.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
# pkg_resources==0.0.0
pooch==1.6.0
proglog==0.1.10
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
pims==0.5
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.13.0
setproctitle==1.3.2
shapely==2.0.0
six==1.16.0
smmap==5.0.0
soundfile==0.11.0
soupsieve==2.3.2.post1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2021.11.2
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
urllib3==1.26.14
wandb==0.13.9
wcwidth==0.2.6
Werkzeug==2.2.2
zipp==3.11.0
torch==1.13.1
